# Refactoring for `pickModel.js`