@javagl 2024-02-16
I mentioned

> The new approach for the intersection tests will transform the ray, and not the vertices. One place in the code is marked with a TODO to point out that the "vertical exaggeration" may have to be taken into account here (I haven't yet thought about the details how this could be done)

and I've accidentally been thinking about this. A relatively straightforward approach could be to apply ... the inverse of the exaggeration (a ... belittling... or floccinaucinihilipilification ... ?) to the ray before transforming it with the actual inverse transform. The difficult point here is that the exaggeration is non-linear (it is linear _along the normal_ (at least, I think that there is no such thing like a 'logarithmic exaggeration', even though that could be interesting as well), but it is not linear on the scale of the globe). As a last resort, there would always be the 'safe fallback' like `if (verticalExaggeration !== 1.0) useTheOldApproach();`...