@ggetz 2024-02-19
@javagl I think the proposed approach could be used with vertical exaggeration if both the ray and the position are transformed into the vertically exaggerated ellipsoid space. However, this will still end up applying a transform to each vertex, which I assume would nullify the optimization.

As you said above, we could do this only when `verticalExaggeration !== 1.0`. While still slower in the vertical exaggeration case, transforming both the ray and position to scaled ellipsoid space will avoid too much branch code.