@ggetz 2024-02-19
Actually, upon further review, that is only possible for geocentric exaggeration, not geodetic exaggeration as vertical exaggeration works elsewhere.