@ggetz 2024-05-10
@javagl I'm going to go ahead and close this PR for now. Some of these improvements were integrated with https://github.com/CesiumGS/cesium/pull/11842, and we'll want to make sure we have the proper time to go through the rest.

Thanks!